Move audio sharing to a separate feature module
