+++ app/build.gradle
@@ -129,8 +129,10 @@ dependencies {
   implementation project(path: ':common:toolbar')
   implementation project(path: ':common:upgrade')
   implementation project(path: ':common:ui:core')
+  implementation project(path: ':common:util')
 
   implementation project(path: ':feature:audio')
+  implementation project(path: ':feature:audioshare')
   implementation project(path: ':feature:downloadmanager')
   implementation project(path: ':feature:qarilist')
 
@@ -170,7 +172,7 @@ dependencies {
   kapt("com.squareup.moshi:moshi-kotlin-codegen:${moshiVersion}")
 
   implementation "dev.chrisbanes:insetter-ktx:0.3.1"
-  implementation 'com.jakewharton.timber:timber:5.0.1'
+  implementation "com.jakewharton.timber:timber:${timberVersion}"
   debugImplementation 'com.squareup.leakcanary:leakcanary-android:2.9.1'
   implementation 'com.google.firebase:firebase-crashlytics:18.2.13'
