+++ app/src/main/java/com/quran/labs/androidquran/database/AudioDatabaseVersionChecker.kt
@@ -1,5 +1,6 @@
 package com.quran.labs.androidquran.database
 
+import com.quran.labs.androidquran.common.audio.timing.SuraTimingDatabaseHandler
 import com.quran.labs.androidquran.feature.audio.VersionableDatabaseChecker
 import javax.inject.Inject
