+++ common/audio/build.gradle
@@ -16,11 +16,13 @@ android {
 dependencies {
   implementation project(":common:data")
   implementation project(":common:download")
+  implementation project(path: ':common:util')
 
   implementation deps.dagger.runtime
   implementation "androidx.annotation:annotation:${androidxAnnotationVersion}"
   implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:${coroutinesVersion}"
   implementation "com.squareup.okio:okio:${okioVersion}"
+  implementation "com.jakewharton.timber:timber:${timberVersion}"
 
   testImplementation "junit:junit:${junitVersion}"
   testImplementation "com.google.truth:truth:${truthVersion}"
