+++ feature/audioshare/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.quran.labs.androidquran.feature.audioshare"/>
