+++ feature/audioshare/src/main/java/com/quran/labs/androidquran/feature/audioshare/soundfile/CheapMP3.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.quran.labs.androidquran.feature.audio.util.soundfile;
+package com.quran.labs.androidquran.feature.audioshare.soundfile;
 
 import java.io.File;
 import java.io.FileInputStream;
