+++ settings.gradle
@@ -11,10 +11,12 @@ include ':common:reading'
 include ':common:recitation'
 include ':common:search'
 include ':common:toolbar'
+include ':common:util'
 include ':common:upgrade'
 include ':common:ui:core'
 include ':feature:analytics-noop'
 include ':feature:audio'
+include ':feature:audioshare'
 include ':feature:downloadmanager'
 include ':feature:qarilist'
 include ':feature:recitation'
